score/requirements.txt
RateMyProfessorAPI==1.3.6
vaderSentiment==3.3.2